Move rumEventsCollection/longTask
