+++ packages/rum-core/src/boot/startRum.ts
@@ -18,7 +18,7 @@ import { startViewContexts } from '../domain/contexts/viewContexts'
 import { startRequestCollection } from '../domain/requestCollection'
 import { startActionCollection } from '../domain/action/actionCollection'
 import { startErrorCollection } from '../domain/error/errorCollection'
-import { startLongTaskCollection } from '../domain/rumEventsCollection/longTask/longTaskCollection'
+import { startLongTaskCollection } from '../domain/longTask/longTaskCollection'
 import { startResourceCollection } from '../domain/rumEventsCollection/resource/resourceCollection'
 import { startViewCollection } from '../domain/rumEventsCollection/view/viewCollection'
 import type { RumSessionManager } from '../domain/rumSessionManager'
