+++ packages/rum-core/src/domain/longTask/longTaskCollection.ts
@@ -1,9 +1,9 @@
 import { toServerDuration, relativeToClocks, generateUUID } from '@datadog/browser-core'
-import type { RawRumLongTaskEvent } from '../../../rawRumEvent.types'
-import { RumEventType } from '../../../rawRumEvent.types'
-import type { LifeCycle } from '../../lifeCycle'
-import { LifeCycleEventType } from '../../lifeCycle'
-import type { RumSessionManager } from '../../rumSessionManager'
+import type { RawRumLongTaskEvent } from '../../rawRumEvent.types'
+import { RumEventType } from '../../rawRumEvent.types'
+import type { LifeCycle } from '../lifeCycle'
+import { LifeCycleEventType } from '../lifeCycle'
+import type { RumSessionManager } from '../rumSessionManager'
 
 export function startLongTaskCollection(lifeCycle: LifeCycle, sessionManager: RumSessionManager) {
   lifeCycle.subscribe(LifeCycleEventType.PERFORMANCE_ENTRIES_COLLECTED, (entries) => {
